whisper_ollama/requirements.txt
openai-whisper
langchain_ollama
langchain

# Dependência externa necessária: FFmpeg (https://ffmpeg.org/download.html)